Extract series test util
